mlops/requirements.txt
google-cloud-aiplatform>=1.36.0
google-cloud-monitoring>=2.15.0
google-cloud-logging>=3.8.0
google-cloud-billing>=1.10.0
pyyaml>=6.0.1
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
plotly>=5.15.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0 